+++ src/main/java/com/likeTrello/board/service/BoardServiceImpl.java
@@ -2,6 +2,8 @@ package com.likeTrello.board.service;
 
 import com.likeTrello.board.model.Board;
 import com.likeTrello.board.repository.BoardRepository;
+import com.likeTrello.exceptions.BoardNotFoundException;
+import com.likeTrello.exceptions.InvalidParameterException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +17,10 @@ public class BoardServiceImpl implements BoardService {
 
     @Override
     public Board getById(Long id) {
+        if (boardRepository.findById(id).isEmpty()) {
+            throw new InvalidParameterException("Invalid board id");
+        }
+
         return boardRepository.findById(id).get();
     }
 
@@ -25,11 +31,19 @@ public class BoardServiceImpl implements BoardService {
 
     @Override
     public void delete(Long id) {
-        boardRepository.deleteById(id);
+        if (boardRepository.findById(id).isEmpty()) {
+            throw new InvalidParameterException("Invalid board id");
+        } else {
+            boardRepository.deleteById(id);
+        }
     }
 
     @Override
     public List<Board> getAll() {
+        if (boardRepository.findAll().isEmpty()) {
+            throw new BoardNotFoundException("No boards exist");
+        }
+
         return boardRepository.findAll();
     }
 }
